+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 ### Fix
 
 - round seconds on manual entry ([#85](https://github.com/peterpeterparker/tietracker/issues/85))
-- timezone date comparison in entry editing ([#87](https://github.com/peterpeterparker/tietracker/issues/87)) 
+- timezone date comparison in entry editing ([#87](https://github.com/peterpeterparker/tietracker/issues/87))
+- native file system breaking changes ([file-system-access/changes.md](https://github.com/WICG/file-system-access/blob/master/changes.md)) 
 
 <a name="1.4.0"></a>